# pkg/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for ExchangeStatus.
const (
	ExchangeStatusCleared     ExchangeStatus = "cleared"
	ExchangeStatusReconciling ExchangeStatus = "reconciling"
	ExchangeStatusUncleared   ExchangeStatus = "uncleared"
)

// Defines values for ServerHealthStatus.
const (
	ServerHealthStatusHealthy   ServerHealthStatus = "healthy"
	ServerHealthStatusUnhealthy ServerHealthStatus = "unhealthy"
)

// Book defines model for Book.
type Book struct {
	CreatedAt              time.Time `json:"created_at"`
	DefaultCurrencyIsoCode string    `json:"default_currency_iso_code"`
	Id                     ID        `json:"id"`
	Name                   string    `json:"name"`
	UpdatedAt              time.Time `json:"updated_at"`
}

// BookProperties defines model for BookProperties.
type BookProperties struct {
	DefaultCurrencyIsoCode string `json:"default_currency_iso_code"`
	Name                   string `json:"name"`
}

// Error A machine-readable format for specifying errors in HTTP API responses
// based on RFC 7807 (https://datatracker.ietf.org/doc/html/rfc7807).
type Error struct {
	// Detail A human-readable explanation specific to this occurrence of the problem.
	Detail *string `json:"detail,omitempty"`

	// Instance A URI reference that identifies the specific occurrence of the problem. It may or may not yield further information if dereferenced.
	Instance *string `json:"instance,omitempty"`

	// Status The HTTP status code generated by the origin server for this occurrence of the problem.
	Status int `json:"status"`

	// Title A short, human-readable summary of the problem type.
	Title string `json:"title"`

	// Type A URI reference that identifies the problem type.
	Type *string `json:"type,omitempty"`
}

// Exchange defines model for Exchange.
type Exchange struct {
	Cheque      *string            `json:"cheque,omitempty"`
	CreatedAt   time.Time          `json:"created_at"`
	Date        openapi_types.Date `json:"date"`
	Description string             `json:"description"`
	Id          ID                 `json:"id"`
	Memo        *string            `json:"memo,omitempty"`
	RegisterId  ID                 `json:"register_id"`
	Status      ExchangeStatus     `json:"status"`
	UpdatedAt   time.Time          `json:"updated_at"`
}

// ExchangeProperties defines model for ExchangeProperties.
type ExchangeProperties struct {
	Cheque      *string            `json:"cheque,omitempty"`
	Date        openapi_types.Date `json:"date"`
	Description string             `json:"description"`
	Memo        *string            `json:"memo,omitempty"`
	RegisterId  ID                 `json:"register_id"`
	Status      ExchangeStatus     `json:"status"`
}

// ExchangeStatus defines model for ExchangeStatus.
type ExchangeStatus string

// ExchangeWithSplits defines model for ExchangeWithSplits.
type ExchangeWithSplits struct {
	Cheque      *string            `json:"cheque,omitempty"`
	CreatedAt   time.Time          `json:"created_at"`
	Date        openapi_types.Date `json:"date"`
	Description string             `json:"description"`
	Id          ID                 `json:"id"`
	Memo        *string            `json:"memo,omitempty"`
	RegisterId  ID                 `json:"register_id"`
	Splits      []Split            `json:"splits"`
	Status      ExchangeStatus     `json:"status"`
	UpdatedAt   time.Time          `json:"updated_at"`
}

// ExistingResource defines model for ExistingResource.
type ExistingResource struct {
	CreatedAt time.Time `json:"created_at"`
	Id        ID        `json:"id"`
	UpdatedAt time.Time `json:"updated_at"`
}

// ID defines model for ID.
type ID = string

// ServerHealthStatus defines model for ServerHealthStatus.
type ServerHealthStatus string

// Split defines model for Split.
type Split struct {
	Amount                int64          `json:"amount"`
	CounterpartAmount     *int64         `json:"counterpart_amount,omitempty"`
	CreatedAt             time.Time      `json:"created_at"`
	DestinationRegisterId ID             `json:"destination_register_id"`
	ExchangeId            ID             `json:"exchange_id"`
	Id                    ID             `json:"id"`
	Memo                  *string        `json:"memo,omitempty"`
	Status                ExchangeStatus `json:"status"`
	UpdatedAt             time.Time      `json:"updated_at"`
}

// SplitProperties defines model for SplitProperties.
type SplitProperties struct {
	Amount                int64          `json:"amount"`
	CounterpartAmount     *int64         `json:"counterpart_amount,omitempty"`
	DestinationRegisterId ID             `json:"destination_register_id"`
	ExchangeId            ID             `json:"exchange_id"`
	Memo                  *string        `json:"memo,omitempty"`
	Status                ExchangeStatus `json:"status"`
}

// Page defines model for page.
type Page = int

// PageSize defines model for page_size.
type PageSize = int

// GetBooksParams defines parameters for GetBooks.
type GetBooksParams struct {
	// Page The page number to retrieve (1-indexed).
	Page *Page `form:"page,omitempty" json:"page,omitempty"`

	// PageSize The maximum number of items to return per page.
	PageSize *PageSize `form:"page-size,omitempty" json:"page-size,omitempty"`
}

// CreateBookJSONBody defines parameters for CreateBook.
type CreateBookJSONBody struct {
	Book BookProperties `json:"book"`
}

// CreateBookJSONRequestBody defines body for CreateBook for application/json ContentType.
type CreateBookJSONRequestBody CreateBookJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /books)
	GetBooks(w http.ResponseWriter, r *http.Request, params GetBooksParams)

	// (POST /books)
	CreateBook(w http.ResponseWriter, r *http.Request)

	// (GET /books/{bookId})
	GetBook(w http.ResponseWriter, r *http.Request, bookId string)

	// (GET /exchanges)
	GetExchanges(w http.ResponseWriter, r *http.Request)

	// (GET /health)
	GetHealth(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// (GET /books)
func (_ Unimplemented) GetBooks(w http.ResponseWriter, r *http.Request, params GetBooksParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /books)
func (_ Unimplemented) CreateBook(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /books/{bookId})
func (_ Unimplemented) GetBook(w http.ResponseWriter, r *http.Request, bookId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /exchanges)
func (_ Unimplemented) GetExchanges(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /health)
func (_ Unimplemented) GetHealth(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetBooks operation middleware
func (siw *ServerInterfaceWrapper) GetBooks(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetBooksParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", r.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "page-size" -------------

	err = runtime.BindQueryParameter("form", true, false, "page-size", r.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page-size", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetBooks(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateBook operation middleware
func (siw *ServerInterfaceWrapper) CreateBook(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateBook(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetBook operation middleware
func (siw *ServerInterfaceWrapper) GetBook(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "bookId" -------------
	var bookId string

	err = runtime.BindStyledParameterWithOptions("simple", "bookId", chi.URLParam(r, "bookId"), &bookId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "bookId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetBook(w, r, bookId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetExchanges operation middleware
func (siw *ServerInterfaceWrapper) GetExchanges(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetExchanges(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/books", wrapper.GetBooks)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/books", wrapper.CreateBook)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/books/{bookId}", wrapper.GetBook)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/exchanges", wrapper.GetExchanges)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/health", wrapper.GetHealth)
	})

	return r
}

type GetBooksRequestObject struct {
	Params GetBooksParams
}

type GetBooksResponseObject interface {
	VisitGetBooksResponse(w http.ResponseWriter) error
}

type GetBooks200JSONResponse struct {
	Books []Book `json:"books"`
}

func (response GetBooks200JSONResponse) VisitGetBooksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type CreateBookRequestObject struct {
	Body *CreateBookJSONRequestBody
}

type CreateBookResponseObject interface {
	VisitCreateBookResponse(w http.ResponseWriter) error
}

type CreateBook201JSONResponse struct {
	Book Book `json:"book"`
}

func (response CreateBook201JSONResponse) VisitCreateBookResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateBook422JSONResponse Error

func (response CreateBook422JSONResponse) VisitCreateBookResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(422)

	return json.NewEncoder(w).Encode(response)
}

type GetBookRequestObject struct {
	BookId string `json:"bookId"`
}

type GetBookResponseObject interface {
	VisitGetBookResponse(w http.ResponseWriter) error
}

type GetBook200JSONResponse struct {
	Book Book `json:"book"`
}

func (response GetBook200JSONResponse) VisitGetBookResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetBook404JSONResponse Error

func (response GetBook404JSONResponse) VisitGetBookResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GetExchangesRequestObject struct {
}

type GetExchangesResponseObject interface {
	VisitGetExchangesResponse(w http.ResponseWriter) error
}

type GetExchanges200JSONResponse struct {
	Exchanges []ExchangeWithSplits `json:"exchanges"`
}

func (response GetExchanges200JSONResponse) VisitGetExchangesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetHealthRequestObject struct {
}

type GetHealthResponseObject interface {
	VisitGetHealthResponse(w http.ResponseWriter) error
}

type GetHealth200JSONResponse struct {
	Reason *string            `json:"reason,omitempty"`
	Status ServerHealthStatus `json:"status"`
}

func (response GetHealth200JSONResponse) VisitGetHealthResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {

	// (GET /books)
	GetBooks(ctx context.Context, request GetBooksRequestObject) (GetBooksResponseObject, error)

	// (POST /books)
	CreateBook(ctx context.Context, request CreateBookRequestObject) (CreateBookResponseObject, error)

	// (GET /books/{bookId})
	GetBook(ctx context.Context, request GetBookRequestObject) (GetBookResponseObject, error)

	// (GET /exchanges)
	GetExchanges(ctx context.Context, request GetExchangesRequestObject) (GetExchangesResponseObject, error)

	// (GET /health)
	GetHealth(ctx context.Context, request GetHealthRequestObject) (GetHealthResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetBooks operation middleware
func (sh *strictHandler) GetBooks(w http.ResponseWriter, r *http.Request, params GetBooksParams) {
	var request GetBooksRequestObject

	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetBooks(ctx, request.(GetBooksRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetBooks")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetBooksResponseObject); ok {
		if err := validResponse.VisitGetBooksResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreateBook operation middleware
func (sh *strictHandler) CreateBook(w http.ResponseWriter, r *http.Request) {
	var request CreateBookRequestObject

	var body CreateBookJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateBook(ctx, request.(CreateBookRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateBook")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateBookResponseObject); ok {
		if err := validResponse.VisitCreateBookResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetBook operation middleware
func (sh *strictHandler) GetBook(w http.ResponseWriter, r *http.Request, bookId string) {
	var request GetBookRequestObject

	request.BookId = bookId

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetBook(ctx, request.(GetBookRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetBook")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetBookResponseObject); ok {
		if err := validResponse.VisitGetBookResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetExchanges operation middleware
func (sh *strictHandler) GetExchanges(w http.ResponseWriter, r *http.Request) {
	var request GetExchangesRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetExchanges(ctx, request.(GetExchangesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetExchanges")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetExchangesResponseObject); ok {
		if err := validResponse.VisitGetExchangesResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetHealth operation middleware
func (sh *strictHandler) GetHealth(w http.ResponseWriter, r *http.Request) {
	var request GetHealthRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetHealth(ctx, request.(GetHealthRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetHealth")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetHealthResponseObject); ok {
		if err := validResponse.VisitGetHealthResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}