# pkg/api/api.openapi.yaml
openapi: "3.0.4"
info:
  title: Finances API
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /books:
    get:
      operationId: GetBooks
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/page_size"
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: object
                required:
                  - books
                properties:
                  books:
                    type: array
                    items:
                      $ref: "#/components/schemas/Book"
    post:
      operationId: CreateBook
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              required:
                - book
              properties:
                book:
                  $ref: "#/components/schemas/BookProperties"
      responses:
        "201":
          description: Created
          content:
            "application/json":
              schema:
                type: object
                required:
                  - book
                properties:
                  book:
                    $ref: "#/components/schemas/Book"
        "422":
          description: Unprocessable Entity
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
  "/books/{bookId}":
    parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: GetBook
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: object
                required:
                  - book
                properties:
                  book:
                    $ref: "#/components/schemas/Book"
        "404":
          description: Not Found
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
  /exchanges:
    get:
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: object
                required:
                  - exchanges
                properties:
                  exchanges:
                    type: array
                    items:
                      $ref: "#/components/schemas/ExchangeWithSplits"
  /health:
    get:
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: object
                properties:
                  status:
                    $ref: "#/components/schemas/ServerHealthStatus"
                  reason:
                    type: string
                required:
                  - status
components:
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: The page number to retrieve (1-indexed).
    page_size:
      name: page-size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50 # if this changes, change "controller.DefaultPageSize" as well.
        default: 20
      description: The maximum number of items to return per page.
  schemas:
    Error:
      type: object
      description: |
        A machine-readable format for specifying errors in HTTP API responses
        based on RFC 7807 (https://datatracker.ietf.org/doc/html/rfc7807).
      required:
        - title
        - status
      properties:
        type:
          type: string
          format: uri-reference
          description: A URI reference that identifies the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          format: uri-reference
          description: A URI reference that identifies the specific occurrence of the problem. It may or may not yield further information if dereferenced.

    ID:
      type: string

    ExistingResource:
      type: object
      required:
        - id
        - created_at
        - updated_at
      properties:
        id:
          $ref: "#/components/schemas/ID"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Book:
      type: object
      allOf:
        - $ref: "#/components/schemas/ExistingResource"
        - $ref: "#/components/schemas/BookProperties"

    BookProperties:
      type: object
      required:
        - name
        - default_currency_iso_code
      properties:
        name:
          type: string
        default_currency_iso_code:
          type: string

    ServerHealthStatus:
      type: string
      enum:
        - healthy
        - unhealthy

    ExchangeStatus:
      type: string
      enum:
        - uncleared
        - reconciling
        - cleared

    ExchangeWithSplits:
      type: object
      allOf:
        - $ref: "#/components/schemas/Exchange"
        - type: object
          required:
            - splits
          properties:
            splits:
              type: array
              items:
                $ref: "#/components/schemas/Split"

    ExchangeProperties:
      type: object
      required:
        - date
        - register_id
        - description
        - status
      properties:
        date:
          type: string
          format: date
        register_id:
          $ref: "#/components/schemas/ID"
        cheque:
          type: string
        description:
          type: string
        memo:
          type: string
        status:
          $ref: "#/components/schemas/ExchangeStatus"

    Exchange:
      allOf:
        - $ref: "#/components/schemas/ExistingResource"
        - $ref: "#/components/schemas/ExchangeProperties"

    Split:
      type: object
      allOf:
        - $ref: "#/components/schemas/ExistingResource"
        - $ref: "#/components/schemas/SplitProperties"

    SplitProperties:
      type: object
      required:
        - exchange_id
        - destination_register_id
        - amount
        - status
      properties:
        exchange_id:
          $ref: "#/components/schemas/ID"
        destination_register_id:
          $ref: "#/components/schemas/ID"
        amount:
          type: integer
          x-go-type: int64
        counterpart_amount:
          type: integer
          x-go-type: int64
        memo:
          type: string
        status:
          $ref: "#/components/schemas/ExchangeStatus"

    UserProperties:
      type: object
      required:
        - display_name
      properties:
        display_name:
          type: string

    User:
      allOf:
        - $ref: "#/components/schemas/ExistingResource"
        - $ref: "#/components/schemas/UserProperties"